schema-001 fixes

+++ DSL/Resql/training/GET/service_management/list_requested_service_triggers.sql
@@ -63,4 +63,4 @@ ORDER BY
     CASE WHEN :sorting = 'serviceName desc' THEN service_name END DESC,
     CASE WHEN :sorting = 'requestedAt asc' THEN created END ASC,
     CASE WHEN :sorting = 'requestedAt desc' THEN created END DESC
-OFFSET ((GREATEST(:page, 1) - 1) * :page_size) LIMIT :page_size;
+OFFSET ((GREATEST(:page::INTEGER, 1) - 1) * :page_size::INTEGER) LIMIT :page_size::INTEGER;
